@@ -1,5 +1,5 @@
 {
-  "name": "gray",
+  "name": "dbex",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -9,28 +9,57 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@auth/prisma-adapter": "^2.2.0",
+    "@monaco-editor/react": "^4.6.0",
+    "@nextui-org/skeleton": "^2.0.29",
+    "@nextui-org/system": "^2.2.2",
+    "@nextui-org/theme": "^2.2.6",
+    "@phosphor-icons/react": "^2.1.6",
+    "@prisma/client": "^5.15.1",
+    "@prisma/extension-accelerate": "^1.1.0",
+    "@radix-ui/react-accordion": "^1.1.2",
+    "@radix-ui/react-alert-dialog": "^1.0.5",
+    "@radix-ui/react-avatar": "^1.0.4",
+    "@radix-ui/react-context-menu": "^2.1.5",
+    "@radix-ui/react-dialog": "^1.1.0",
+    "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-icons": "^1.3.0",
+    "@radix-ui/react-select": "^2.0.0",
+    "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.0.2",
+    "@radix-ui/react-tabs": "^1.0.4",
+    "@radix-ui/react-tooltip": "^1.0.7",
+    "@uiw/react-textarea-code-editor": "^3.0.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "cmdk": "^1.0.0",
+    "emoji-picker-react": "^4.10.0",
     "framer-motion": "^11.2.10",
-    "mini-svg-data-uri": "^1.4.4",
-    "next": "14.2.4",
+    "next": "14.2.1",
     "next-auth": "5.0.0-beta.19",
-    "react": "^18",
-    "react-dom": "^18",
+    "next-themes": "^0.3.0",
+    "papaparse": "^5.4.1",
+    "pg": "^8.12.0",
+    "phospho": "1.3.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-resizable-panels": "^2.0.19",
+    "sonner": "^1.5.0",
     "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.4",
-    "mini-css-extract-plugin": "^2.9.0",
-    "postcss": "^8",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "@types/node": "^20.14.2",
+    "@types/papaparse": "^5.3.14",
+    "@types/pg": "^8.11.6",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "autoprefixer": "^10.4.19",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "14.2.1",
+    "postcss": "^8.4.38",
+    "prisma": "^5.15.1",
+    "tailwindcss": "^3.4.4",
+    "typescript": "^5.4.5"
   }
 }
