@@ -1,6 +1,10 @@
 import NextAuth from "next-auth"
 import GitHub from "next-auth/providers/github"
+import Google from "next-auth/providers/google"
+import { PrismaAdapter } from "@auth/prisma-adapter"
+import { prisma } from "./lib/prisma"
 
 export const { handlers, auth, signIn, signOut } = NextAuth({
-  providers: [GitHub],
+  adapter: PrismaAdapter(prisma),
+  providers: [GitHub, Google],
 })
